+++ lua/codecompanion/config.lua
@@ -925,6 +925,7 @@ You must create or modify a workspace file through a series of prompts over mult
         height = 0.8,
         width = 0.45,
         relative = "editor",
+        full_height = true,
         opts = {
           breakindent = true,
           cursorcolumn = false,
@@ -938,6 +939,7 @@ You must create or modify a workspace file through a series of prompts over mult
           wrap = true,
         },
       },
+      auto_scroll = true, -- Automatically scroll down and place the cursor at the end
       intro_message = "Welcome to CodeCompanion ✨! Press ? for options",
 
       show_header_separator = false, -- Show header separators in the chat buffer? Set this to false if you're using an external markdown formatting plugin
@@ -1013,6 +1015,7 @@ You must:
 - Avoid including line numbers in code blocks.
 - Avoid wrapping the whole response in triple backticks.
 - Only return code that's directly relevant to the task at hand. You may omit code that isn’t necessary for the solution.
+- Avoid using H1 and H2 headers in your responses.
 - Use actual line breaks in your responses; only use "\n" when you want a literal backslash followed by 'n'.
 - All non-code text responses must be written in the %s language indicated.