+++ lua/codecompanion/strategies/chat/agents/tools/editor.lua
@@ -103,7 +103,7 @@ return {
     ---@param self CodeCompanion.Agent.Tool The Tools object
     ---@param actions table The action object
     ---@param input? any The output from the previous function call
-    ---@return nil|{ status: string, msg: string }
+    ---@return nil|{ status: "success"|"error", data: any }
     function(self, actions, input)
       ---Run the action
       ---@param action table