+++ lua/codecompanion/strategies/cmd.lua
@@ -48,13 +48,14 @@ function Cmd:start()
   client.new({ adapter = self.adapter:map_schema_to_params() }):request(self.adapter:map_roles(messages), {
     ---@param err string
     ---@param data table
-    callback = function(err, data)
+    ---@param adapter CodeCompanion.Adapter The modified adapter from the http client
+    callback = function(err, data, adapter)
       if err then
         return log:error(err)
       end
 
       if data then
-        local result = self.adapter.handlers.chat_output(self.adapter, data)
+        local result = self.adapter.handlers.chat_output(adapter, data)
         if result and result.output and result.output.content then
           local content = result.output.content
           content:gsub("^%s*(.-)%s*$", "%1")