+++ lua/codecompanion/strategies/init.lua
@@ -123,7 +123,7 @@ function Strategies:chat()
         return create_chat(opts.user_prompt)
       end
 
-      vim.ui.input({
+      return vim.ui.input({
         prompt = string.gsub(self.context.filetype, "^%l", string.upper) .. " " .. config.display.action_palette.prompt,
       }, function(input)
         if not input then
@@ -256,6 +256,9 @@ function Strategies.evaluate_prompts(prompts, context)
     end)
     :map(function(prompt)
       local content = type(prompt.content) == "function" and prompt.content(context) or prompt.content
+      if prompt.role == config.constants.SYSTEM_ROLE and not prompt.opts then
+        prompt.opts = { visible = false }
+      end
       return {
         role = prompt.role or "",
         content = content,